requirements.txt
asgiref==3.8.1
Django==5.0.4
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pep8-naming==0.13.3
phonenumberslite==8.13.35
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
